trying to add graphql and mysql
