+++ server/graphql/schemas/mutations.schema.js
@@ -1,0 +1,5 @@
+module.exports = `
+type Mutation {
+    createItem(name: String!): Item!
+    deleteItem(id: ID!): Boolean!
+  }`;
