+++ server/graphql/schemas/query.schema.js
@@ -1,0 +1,9 @@
+module.exports = `
+type Query {
+    me: User
+    user(id: ID!): User
+    users: [User!]
+
+    item(id: ID!): Item
+    items: [Item!]
+}`;
