+++ server/graphql/schemas/types.schema.js
@@ -1,0 +1,12 @@
+module.exports = `
+type User {
+    id: ID!
+    username: String!
+    item: [Item!]
+  }
+
+  type Item {
+    id: ID!
+    name: String!
+    user: User!
+  }`;
