+++ server/models/index.js
@@ -26,11 +26,12 @@ fs
   })
   .forEach(file => {
     // var model = sequelize['import'](path.join(__dirname, file));
-    var model = require(path.join(__dirname, file))(sequelize, Sequelize.DataTypes);
+    const model = require(path.join(__dirname, file))(sequelize, Sequelize.DataTypes);
     db[model.name] = model;
   });
 
 Object.keys(db).forEach(modelName => {
+  console.log(modelName);
   if (db[modelName].associate) {
     db[modelName].associate(db);
   }
