+++ server/models/item.js
@@ -2,15 +2,9 @@
 module.exports = (sequelize, DataTypes) => {
 
 const Item = sequelize.define('item', {
-  id: 
-    {
-      type: DataTypes.UUID,
-      primaryKey: true,
-      defaultValue: sequelize.UUIDV4 
-    },
-  item_name: DataTypes.STRING,
-  desc: DataTypes.STRING,
-  price: DataTypes.INTEGER,
+  name: {
+      type: DataTypes.STRING,
+    }
 })
 
 Item.associate = (models) => {
